+++ src/main/java/com/supermartijn642/wormhole/portal/IPortalGroupEntity.java
@@ -9,6 +9,4 @@ public interface IPortalGroupEntity {
 
     PortalGroup getGroup();
 
-    void onBreak();
-
 }